퀘스트 창, NPC 퀘스트 선택창 구현
퀘스트 중복, 이미 완료한 퀘스트인지 확인, 퀘스트 보상 구현,

+++ Assets/03_Scripts/UI/ScriptableObject/Quest/Quest.cs
@@ -21,7 +21,7 @@ public class Quest
 
     public float GetProgress()
     {
-        return (m_iCurrentAmount / m_iAmount) * 100.0f;
+        return ((float)m_iCurrentAmount / m_iAmount) * 100.0f;
     }
 
     public bool UpdateProgress(int _iAmount)
